src/requirements.txt
kubernetes==26.1.0
requests==2.31.0
python-json-logger==2.0.7
logfmter==0.0.6
grafana-client==3.5.0
gitpython==3.1.32
sqlalchemy==2.0.19
